+++ paper/paper.bib
@@ -38,6 +38,7 @@
   publisher = {Zenodo},
   journal   = {GitHub repository},
   url       = {https://github.com/pangeo-data/xESMF}
+  doi       = {10.5281/zenodo.4294774}
 }
 
 @misc{tasmantides,
@@ -187,6 +188,7 @@
   eprint        = {2309.06662},
   archivePrefix = {arXiv},
   primaryClass  = {physics.ao-ph}
+  doi           = {10.48550/arXiv.2309.06662}
 }
 
 @incollection{Gula2022ch8,
