+++ paper/paper.bib
@@ -37,7 +37,7 @@
   year      = {2023},
   publisher = {Zenodo},
   journal   = {GitHub repository},
-  url       = {https://github.com/pangeo-data/xESMF}
+  url       = {https://github.com/pangeo-data/xESMF},
   doi       = {10.5281/zenodo.4294774}
 }
 
@@ -187,7 +187,7 @@
   year          = {2023},
   eprint        = {2309.06662},
   archivePrefix = {arXiv},
-  primaryClass  = {physics.ao-ph}
+  primaryClass  = {physics.ao-ph},
   doi           = {10.48550/arXiv.2309.06662}
 }
 
