### 2024-04-27 #6686 @navidcy
@kthyng, regarding reviewers we can suggest:

Gustavo Marques @ gustavo-marques
Frederic Castruccio @ fredc
Alper Altuntas @ alperaltuntas
Michael Levy @ mnlevy1981
Theresa Morrison @ theresa-morrison
Marshall Ward @ marshallward (possible conflict? one of the paper's coauthors has written a paper with them on 2020)
Guilherme P. Castelao  @ castelao
Dion Häfner @ dionhaefner
Keith Lindsay @ klindsay28

I couldn't find the JOSS conflict policy; could you direct us to that?
