### 2024-04-27 305f169
add JOSS badge (#169)
