+++ paper/paper.bib
@@ -244,7 +244,8 @@
   author    = {Richardson, L. F.},
   publisher = {Cambridge University Press},
   address   = {London, UK},
-  year      = {1922}
+  year      = {1922},
+  doi       = {10.1017/CBO9780511618291}
 }
 
 @article{bleck2002gvc,
