### 2024-04-27 #6686 @navidcy
We added 2 missing DOIs.
The rest, namely:

> ```
> MISSING DOIs
>
> - No DOI given, and none found for title: tasman-tides
> - No DOI given, and none found for title: MITgcm_python
> - No DOI given, and none found for title: Pyroms
> - No DOI given, and none found for title: Earth System Modeling Group (ESMG) gridtools
> - No DOI given, and none found for title: Elements of the modular ocean model (MOM)
> - No DOI given, and none found for title: CEFI-regional-MOM6: Essential tools, XML files, an...
>
> ```

are citations to either GitHub repos and therefore do not have a DOI associated with them or to a unpublished MOM5 manual.
