+++ paper/paper.bib
@@ -229,14 +229,14 @@
 
 
 @article{Orlanski1976,
-  author = {I. Orlanski},
-  doi = {10.1016/0021-9991(76)90023-1},
+  author  = {I. Orlanski},
+  doi     = {10.1016/0021-9991(76)90023-1},
   journal = {Journal of Computational Physics},
-  number = {3},
-  pages = {251-269},
-  title = {A simple boundary condition for unbounded hyperbolic flows},
-  volume = {21},
-  year = {1976},
+  number  = {3},
+  pages   = {251-269},
+  title   = {A simple boundary condition for unbounded hyperbolic flows},
+  volume  = {21},
+  year    = {1976},
 }
 
 @book{richardson1922weather,
@@ -278,4 +278,3 @@
   journal   = {GitHub repository},
   url       = {https://github.com/NOAA-GFDL/CEFI-regional-MOM6}
 }
-
