### 2024-04-30 53dcec0
fix renaming typo and combine all outputs into single year rather than separate files
