### 2024-04-30 e16340d
Era5 fix. Typo, and need to combine years into single file again (#170)
* fix renaming typo and combine all outputs into single year rather than separate files

* fix typo and combine all years of era5 into one file

* black
