### 2024-04-30 #6686 @kthyng
@navidcy

> I couldn't find the JOSS conflict policy; could you direct us to that?

https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy

> the repo title doesn't match the paper title. There is a v missing at the end:
Do you want to change the paper title or pre-review issue? I can do that for you if so.
