### 2024-05-06 #6686 @AnjaliSandip
👋 @gustavo-marques, @fredc, @alperaltuntas & @mnlevy1981,  would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.
