### 2024-05-09 d8b3197
attempt to simplify IC
