### 2024-05-14 9005154
Added caption details for nested figure
