+++ paper/paper.bib
@@ -278,3 +278,24 @@
   journal   = {GitHub repository},
   url       = {https://github.com/NOAA-GFDL/CEFI-regional-MOM6}
 }
+
+@article{ACCESS-OM2,
+  author  = {Kiss, A. E. and Hogg, A. McC. and Hannah, N. and Boeira Dias, F. and Brassington, G. B. and Chamberlain, M. A. and Chapman, C. and Dobrohotoff, P. and Domingues, C. M. and Duran, E. R. and England, M. H. and Fiedler, R. and Griffies, S. M. and Heerdegen, A. and Heil, P. and Holmes, R. M. and Klocker, A. and Marsland, S. J. and Morrison, A. K. and Munroe, J. and Nikurashin, M. and Oke, P. R. and Pilo, G. S. and Richet, O. and Savita, A. and Spence, P. and Stewart, K. D. and Ward, M. L. and Wu, F. and Zhang, X.},
+  doi     = {10.5194/gmd-13-401-2020},
+  journal = {Geoscientific Model Development},
+  number  = {2},
+  pages   = {401--442},
+  title   = {{ACCESS-OM2 v1.0}: a global ocean--sea ice model at three resolutions},
+  volume  = {13},
+  year    = {2020}
+}
+
+@article{Tsujino2018,
+  author  = {Hiroyuki Tsujino and Shogo Urakawa and Hideyuki Nakano and R. Justin Small and Who M. Kim and Stephen G. Yeager and Gokhan Danabasoglu and Tatsuo Suzuki and Jonathan L. Bamber and Mats Bentsen and Claus W. B{\"o}ning and Alexandra Bozec and Eric P. Chassignet and Enrique Curchitser and Fabio {Boeira Dias} and Paul J. Durack and Stephen M. Griffies and Yayoi Harada and Mehmet Ilicak and Simon A. Josey and Chiaki Kobayashi and Shinya Kobayashi and Yoshiki Komuro and William G. Large and Julien {Le Sommer} and Simon J. Marsland and Simona Masina and Markus Scheinert and Hiroyuki Tomita and Maria Valdivieso and Dai Yamazaki},
+  doi     = {10.1016/j.ocemod.2018.07.002},
+  journal = {Ocean Modelling},
+  pages   = {79-139},
+  title   = {{JRA-55} based surface dataset for driving ocean--sea-ice models ({JRA55}-do)},
+  volume  = {130},
+  year    = {2018},
+}
