+++ paper/paper.md
@@ -31,6 +31,10 @@ authors:
     orcid: 0000-0002-1222-375X
     affiliation: "1, 2"
     equal-contrib: true
+  - name: Luwei Yang
+    orcid: 0000-0001-8570-7424
+    affiliation: "1"
+    equal-contrib: true
 affiliations:
  - name: Australian National University, Australia
    index: 1
@@ -75,7 +79,7 @@ Pertinent for our discussion, MOM6 provides support for open boundary conditions
 However, setting up a regional configuration for MOM6 can be challenging, time consuming, and often involves using several programming languages, a few different tools, and also manually editing/tweaking some input files.
 The `regional-mom6` package overcomes these difficulties, automatically generating a regional MOM6 configuration of the user's choice with relatively simple domain geometry, that is, rectangular domains.
 
-![A snapshot of the surface speed from a two-tier (one-way) nested regional ocean simulation of the East Australian Current system using MOM6. The large regional model (outer dashed box), named SOUTHPAC-01 is forced at the open boundaries by the global 10th degree model, ACCESS-OM2-01, while the small regional model (inner dashed box), named EAC-003 is forced at the open boundaries by SOUTHPAC-01. Horizontal resolutions are 10th degree for SOUTHPAC-01 and 30th degree for EAC-003, while vertical resolutions are 75 and 100 levels, respectively. All simulations share a common interannual atmospheric forcing, JRA55-do, spanning 1990 to 2018.\label{fig:tasman}](nested-regional-mom6.pdf){ width=90% }
+![Snapshot of the ocean surface speed from a two-tier, one-way nested regional ocean configurations of the East Australian Current. The outer regional configuration uses 1/10th degree horizontal resolution, 75 vertical levels, and is forced by the output from the global ocean--sea ice model at 1/10th degree horizontal resolution (ACCESS-OM2-01; see @ACCESS-OM2). The inner regional configuration uses 1/30th degree horizontal resolution, 100 vertical levels, and is forced with the outer regional model. All simulations share a common inter-annual atmospheric forcing from 1990 to 2018 provided by the JRA55-do reanalysis [@Tsujino2018].\label{fig:tasman}](nested-regional-mom6.pdf){ width=90% }
 
 ## `regional-mom6`
 
