+++ paper/paper.md
@@ -79,7 +79,7 @@ Pertinent for our discussion, MOM6 provides support for open boundary conditions
 However, setting up a regional configuration for MOM6 can be challenging, time consuming, and often involves using several programming languages, a few different tools, and also manually editing/tweaking some input files.
 The `regional-mom6` package overcomes these difficulties, automatically generating a regional MOM6 configuration of the user's choice with relatively simple domain geometry, that is, rectangular domains.
 
-![Snapshot of the ocean surface speed from a two-tier, one-way nested regional ocean configurations of the East Australian Current. The outer regional configuration uses 1/10th degree horizontal resolution, 75 vertical levels, and is forced by the output from the global ocean--sea ice model at 1/10th degree horizontal resolution (ACCESS-OM2-01; see @ACCESS-OM2). The inner regional configuration uses 1/30th degree horizontal resolution, 100 vertical levels, and is forced with the outer regional model. All simulations share a common inter-annual atmospheric forcing from 1990 to 2018 provided by the JRA55-do reanalysis [@Tsujino2018].\label{fig:tasman}](nested-regional-mom6.pdf){ width=90% }
+![Snapshot of the ocean surface speed from a two-tier, one-way nested regional ocean configurations of the East Australian Current. The outer regional configuration uses 1/10ᵒ horizontal resolution, 75 vertical levels, and is forced by the output from the global ocean--sea ice model at 1/10ᵒ horizontal resolution (ACCESS-OM2-01; @ACCESS-OM2). The inner regional configuration uses 1/30ᵒ horizontal resolution, 100 vertical levels, and is forced with the outer regional model. All simulations share a common inter-annual atmospheric forcing from 1990 to 2018 provided by the JRA55-do reanalysis [@Tsujino2018].\label{fig:tasman}](nested-regional-mom6.pdf){ width=90% }
 
 ## `regional-mom6`
 
