### 2024-05-14 857b0a8
add CLEx affiliation + CoE grants
