+++ paper/paper.md
@@ -33,7 +33,7 @@ authors:
     equal-contrib: true
   - name: Luwei Yang
     orcid: 0000-0001-8570-7424
-    affiliation: "1"
+    affiliation: "1, 2"
     equal-contrib: true
 affiliations:
  - name: Australian National University, Australia
@@ -126,7 +126,7 @@ With the technically-challenging aspects of setting up a regional configuration 
 # Acknowledgements
 
 We thank the vibrant community of the Consortium for Ocean–Sea Ice Modeling in Australia ([cosima.org.au](https://cosima.org.au)) and also Josué Martínez-Moreno and Callum Shakespeare for useful discussions during the development of this package.
-We acknowledge support from the Australian Research Council under DECRA Fellowship DE210100749 (N.C.C.) and grant LP200100406 (A.E.K.).
+We acknowledge support from the Australian Research Council under DECRA Fellowship DE210100749 (N.C.C.), grant LP200100406 (A.E.K.), the Center of Excellences for Climate Extremes CE170100023 and for the Weather of the 21st Century CE230100012.
 We would also like to acknowledge the code and notes by James Simkins, Andrew Ross, and Rob Cermak, which helped us to troubleshoot and improve the algorithms in our package.
 
 # References
