### 2024-05-14 8ec1dc4
differentiate inner-outer regions
