### 2024-05-14 4115c76
numpy.min -> numpy.nanmin to avoid issues with masked input
