### 2024-05-14 e9ec8f8
add tests for K->C conversion
