+++ tests/test_expt_class.py
@@ -100,6 +100,88 @@ def test_setup_bathymetry(
     bathymetry_file.unlink()
 
 
+def number_of_gridpoints(longitude_extent, latitude_extent, resolution):
+    nx = int((longitude_extent[-1] - longitude_extent[0]) / resolution)
+    ny = int((latitude_extent[-1] - latitude_extent[0]) / resolution)
+
+    return nx, ny
+
+
+def generate_temperature_arrays(nx, ny, number_vertical_layers):
+
+    temp_in_C = np.random.randn(ny, nx, number_vertical_layers)
+
+    temp_in_C_masked = np.copy(temp_in_C)
+    temp_in_C_masked[20:24, 32:35, :] = float("nan")
+
+    temp_in_K = np.copy(temp_in_C) + 273.15
+    temp_in_K_masked = np.copy(temp_in_C_masked) + 273.15
+
+    if np.nanmin(temp_in_C_masked) == np.min(temp_in_C):
+        return temp_in_C, temp_in_C_masked, temp_in_K, temp_in_K_masked
+    else:
+        return generate_temperature_arrays(nx, ny, number_vertical_layers)
+
+
+def generate_silly_coords(
+    longitude_extent, latitude_extent, resolution, depth, number_vertical_layers
+):
+    nx, ny = number_of_gridpoints(longitude_extent, latitude_extent, resolution)
+
+    horizontal_buffer = 5
+
+    silly_lat = np.linspace(
+        latitude_extent[0] - horizontal_buffer,
+        latitude_extent[1] + horizontal_buffer,
+        ny,
+    )
+    silly_lon = np.linspace(
+        longitude_extent[0] - horizontal_buffer,
+        longitude_extent[1] + horizontal_buffer,
+        nx,
+    )
+    silly_depth = np.linspace(0, depth, number_vertical_layers)
+
+    return silly_lat, silly_lon, silly_depth
+
+
+longitude_extent = [-5, 3]
+latitude_extent = (0, 10)
+date_range = ["2003-01-01 00:00:00", "2003-01-01 00:00:00"]
+resolution = 0.1
+number_vertical_layers = 5
+layer_thickness_ratio = 1
+depth = 1000
+
+silly_lat, silly_lon, silly_depth = generate_silly_coords(
+    longitude_extent, latitude_extent, resolution, depth, number_vertical_layers
+)
+
+dims = ["silly_lat", "silly_lon", "silly_depth"]
+
+coords = {"silly_lat": silly_lat, "silly_lon": silly_lon, "silly_depth": silly_depth}
+
+mom_run_dir = "rundir/"
+mom_input_dir = "inputdir/"
+toolpath_dir = "toolpath"
+grid_type = "even_spacing"
+
+nx, ny = number_of_gridpoints(longitude_extent, latitude_extent, resolution)
+
+temp_in_C, temp_in_C_masked, temp_in_K, temp_in_K_masked = generate_temperature_arrays(
+    nx, ny, number_vertical_layers
+)
+
+temp_C = xr.DataArray(temp_in_C, dims=dims, coords=coords)
+temp_K = xr.DataArray(temp_in_K, dims=dims, coords=coords)
+temp_C_masked = xr.DataArray(temp_in_C_masked, dims=dims, coords=coords)
+temp_K_masked = xr.DataArray(temp_in_K_masked, dims=dims, coords=coords)
+
+
+@pytest.mark.parametrize(
+    "temp_dataarray_initial_condition",
+    [temp_C, temp_C_masked, temp_K, temp_K_masked],
+)
 @pytest.mark.parametrize(
     (
         "longitude_extent",
@@ -116,13 +198,13 @@ def test_setup_bathymetry(
     ),
     [
         (
-            [-5, 5],
-            (0, 10),
[…]
