### 2024-05-14 051da87
don't hardcode mask
