### 2024-05-14 3e26e2d
add another test for temperatures
