### 2024-05-14 edc29ee
convert K->C before removing mask
