### 2024-05-14 ccef393
more robust K -> C
