+++ regional_mom6/regional_mom6.py
@@ -616,10 +616,10 @@ class experiment:
                 "Error in reading in initial condition tracers. Terminating!"
             )
 
-        ## if min(temp) > 100 then assume that units must be degrees K
+        ## if min(temperature) > 100 then assume that units must be degrees K
         ## (otherwise we can't be on Earth) and convert to degrees C
-        if np.nanmin(ic_raw["temp"]) > 100:
-            ic_raw["temp"] -= 273.15
+        if np.nanmin(ic_raw[varnames["tracers"]["temp"]]) > 100:
+            ic_raw[varnames["tracers"]["temp"]] -= 273.15
 
         # Rename all coordinates to have 'lon' and 'lat' to work with the xesmf regridder
         if arakawa_grid == "A":
