### 2024-06-04 #6686 @AnjaliSandip
👋 @theresa-morrison, @castelao, @dionhaefner, and @klindsay28, would you be willing to review this submission for JOSS? We conduct our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.
