### 2024-06-07 #6857 @editorialbot
[REVIEW]: regional-mom6: A Python package for automatic generation of regional configurations for the Modular Ocean Model 6
**Submitting author:** @ashjbarnes (<a href="http://orcid.org/0000-0003-3165-8676">Ashley Barnes</a>)
**Repository:** https://github.com/COSIMA/regional-mom6/
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.6.0
