### 2024-06-11 671be1e
Update inaccurate comment in initial condition generation
