### 2024-06-11 dcca341
Fixes K->C conversion in `initial_condition` (#176)
* numpy.min -> numpy.nanmin to avoid issues with masked input

* add tests for K->C conversion

* don't hardcode mask

* format black

* add clarifying comments

* add another test for temperatures

* format black

* convert K->C before removing mask

* more robust K -> C

* Update inaccurate comment in initial condition generation

* black

* update temp units when they're converted

---------

Co-authored-by: ashjbarnes <ashjbarnes97@gmail.com>
