### 2024-06-11 d025138
check that premade rundir folder exists. If not, try second option, if still not raise valueerror
