### 2024-06-11 187bee5
Force setup_run_directory function to raise AttributeError if it doesn't have access to a cpu layout rather than throw generic error
