### 2024-06-11 5ebb005
add a wrapper function that iterates over four simple boundaries
