### 2024-06-11 7d850f8
additional error handling for new method
