### 2024-06-11 172c3ae
update test function to match
