+++ tests/test_expt_class.py
@@ -471,4 +471,4 @@ def test_rectangular_boundaries(
         "tracers": {"temp": "temp", "salt": "salt"},
     }
 
-    expt.rectangular_boundaries(tmp_path, varnames,["east"])
+    expt.rectangular_boundaries(tmp_path, varnames, ["east"])
