### 2024-06-11 46791e3
include .nc extension to test
