### 2024-07-01 e280646
Update doctest for numpy 2.0.0
