### 2024-07-01 e806ef0
Remove Python 3.11 from testing
For some reason it seems to hang indefinitely. We'll just disable it
for the moment (3.12 works, anyway).
