### 2024-07-01 26ff5fc
Make setup rundir more robust take #2 (#179)
* check that premade rundir folder exists. If not, try second option, if still not raise valueerror

* typo

* Force setup_run_directory function to raise AttributeError if it doesn't have access to a cpu layout rather than throw generic error

* black
