### 2024-07-02 4781919
(v0.6.0) Simplify boundary condition construction (#175)
* attempt to simplify IC

* add a wrapper function that iterates over four simple boundaries

* black

* additional error handling for new method

* update test function to match

* only test east

* black

* include .nc extension to test

---------

Co-authored-by: ashjbarnes <ashjbarnes97@gmail.com>
Co-authored-by: Ashley Barnes <53282288+ashjbarnes@users.noreply.github.com>
