### 2024-07-09 #6857 @dionhaefner
I couldn't test the software myself since I don't have a MOM6 setup, but looking at the tests and CI setup gave me enough confidence to conclude that the project is in good working order. Perhaps the authors want to consider adding integration via [binder](https://mybinder.org/) to allow for easier experimentation with the package.

Based on that I have no notes; looks like a really useful package with sound software engineering and UX. I'm not the biggest fan of requiring conda but I can see how that may be the path of least resistance for the target audience. The need and capabilities of the software are well described in both docs and paper. I congratulate the authors on creating a great piece of scientific software.
