### 2024-07-10 #6857 @alperaltuntas
I have reviewed all the items on my checklist and confirm that the submission is acceptable for publication. My only suggestion for the authors is to consider eliminating the dependency on FRE-NCtools. This dependency complicates the build process unnecessarily. The authors use only three tools from FRE-NCtools, and I believe these functionalities can be implemented in Python within the regional-mom6 package with little to moderate effort. This change would result in a more complete package and a simpler installation process.
