### 2024-07-12 #6857 @navidcy
> My only suggestion for the authors is to consider eliminating the dependency on FRE-NCtools. This dependency complicates the build process unnecessarily. The authors use only three tools from FRE-NCtools, and I believe these functionalities can be implemented in Python within the regional-mom6 package with little to moderate effort. This change would result in a more complete package and a simpler installation process.

Thanks for your comments! @ashjbarnes should we make 3 issues in http://github.com/COSIMA/regional-mom6/issues for these 3 functionalities that @alperaltuntas is referring to?
