### 2024-07-12 #6857 @ashjbarnes
[This issue](https://github.com/COSIMA/regional-mom6/issues/69) is about incorporating the Grid Gen tools. I'll update the description to mention the replacement of FRE-NCtools. I think these two things would have to be done together in one PR, but up to you if we split into separate issues.
