### 2024-07-18 #6857 @ashjbarnes
> > I have reviewed all the items on my checklist and confirm that the submission is acceptable for publication. My only suggestion for the authors is to consider eliminating the dependency on FRE-NCtools. This dependency complicates the build process unnecessarily. The authors use only three tools from FRE-NCtools, and I believe these functionalities can be implemented in Python within the regional-mom6 package with little to moderate effort. This change would result in a more complete package and a simpler installation process.
>
> Thanks @alperaltuntas I agree that this would be good to do. I believe there are some functions that could help us in the [ESMG gridtools repo](https://github.com/ESMG/gridtools/blob/main/source/guides/FRENCtools.rst). We've been meaning for a while to include this repo as a dependency (and acknowledging appropriately) and wrap some of the python grid generation functions to expand the functionality of `regional-mom6`. Extra dependencies might also complicate building and maintainability though of course! What are your thoughts on this?

@AnjaliSandip Please see my response here. I'd be interested in feedback as to whether we should add this existing Python library as a dependency, or write our own NC-FREtools in Python.
