### 2024-07-24 #6857 @alperaltuntas
> @alperaltuntas In your opinion, is the submission acceptable for publication?

@AnjaliSandip Yes, I approve.

> I'd be interested in feedback as to whether we should add this existing Python library as a dependency, or write our own FRE-NCtools in Python.

@ashjbarnes I am unfamiliar with `gridtools`, but sounds like a good alternative.
