### 2024-07-31 #6857 @AnjaliSandip
> Hi @AnjaliSandip, could you give us a bit more guidance in what's needed here? From what I understand @alperaltuntas raised a potential future improvement of the package but they _do_ acknowledge that the software is acceptable for publication as is.
>
> We are very happy to work on this now if you think it's required. Could you please clarify if this is the case and we'll get to it as soon as possible.
All reviewers agree it's ready for publication.
