### 2024-07-31 #6857 @navidcy
Hi @AnjaliSandip. Regarding the missing DOIs above, they are all related to citations to GitHub repos (that have no doi associated) or to the MOM5 manual by Stephen Griffies (again, no doi is associated with that).
