+++ paper/paper.bib
@@ -299,3 +299,11 @@
   volume  = {130},
   year    = {2018},
 }
+
+@article{Silvestri2024GPU,
+  author  = {Simone Silvestri and Gregory L. Wagner and Navid C Constantinou and Christopher Hill and Jean-Michel Campin and Andre Nogueira Souza and Siddhartha Bishnu and Valentin Churavy and John C Marshall and Raffaele Ferrari},
+  title   = {A {GPU}-based ocean dynamical core for routine mesoscale-resolving climate simulations},
+  journal = {ESS Open Archive},
+  year    = {2024},
+  doi     = {10.22541/essoar.171708158.82342448/v1},
+}
