### 2024-08-02 77030b9
Update paper.bib
remove duplicate reference
