+++ paper/paper.bib
@@ -259,16 +259,6 @@
   year    = {2002}
 }
 
-@article{polton2023reproducible,
-  author  = {Polton, J. and Harle, J. and Holt, J. and Katavouta, A. and Partridge, D. and Jardine, J. and Wakelin, S. and Rulent, J. and Wise, A. and Hutchinson, K. and Byrne, D. and Bruciaferri, D. and O'Dea, E. and De Dominicis, M. and Mathiot, P. and Coward, A. and Yool, A. and Palmi\'eri, J. and Lessin, G. and Mayorga-Adame, C. G. and Le Guennec, V. and Arnold, A. and Rousset, C.},
-  doi     = {10.5194/gmd-16-1481-2023},
-  journal = {Geoscientific Model Development},
-  number  = {5},
-  pages   = {1481--1510},
-  title   = {Reproducible and relocatable regional ocean modelling: fundamentals and practices},
-  volume  = {16},
-  year    = {2023}
-}
 
 @misc{CEFIregional,
   author    = {Yi-Cheng Teng and Andrew Ross and Theresa Morrison},
