### 2024-08-02 #6857 @ashjbarnes
Oops I just removed the duplicate reference from the bib file
