### 2024-08-07 #6857 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
