+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'regional-mom6: A Python package for automatic generation of regional configurations for the Modular Ocean Model v6'
+title: 'regional-mom6: A Python package for automatic generation of regional configurations for the Modular Ocean Model 6'
 
 tags:
   - Python
@@ -49,7 +49,7 @@ affiliations:
  - name: University of Tasmania, Australia
    index: 6
 
-date: 26 April 2024
+date: 8 August 2024
 bibliography: paper.bib
 ---
 
@@ -127,6 +127,6 @@ With the technically-challenging aspects of setting up a regional configuration 
 
 We thank the vibrant community of the Consortium for Ocean–Sea Ice Modeling in Australia ([cosima.org.au](https://cosima.org.au)) and also Josué Martínez-Moreno and Callum Shakespeare for useful discussions during the development of this package.
 We acknowledge support from the Australian Research Council under DECRA Fellowship DE210100749 (N.C.C.), grant LP200100406 (A.E.K.), and the Center of Excellences for Climate Extremes CE170100023 and for the Weather of the 21st Century CE230100012.
-We would also like to acknowledge the code and notes by James Simkins, Andrew Ross, and Rob Cermak, which helped us to troubleshoot and improve the algorithms in our package.
+We would also like to acknowledge the code and notes by James Simkins, Andrew Ross, and Rob Cermak, which helped us to troubleshoot and improve the algorithms in our package. Last, we thank the three reviewers and the editor for their efforts and remarks.
 
 # References
