### 2024-08-07 #6857 @navidcy
I've read the proof and it looks good!
