### 2024-08-09 #6857 @kthyng
Note one box isn't checked for @theresa-morrison but as per https://github.com/openjournals/joss-reviews/issues/6857#issuecomment-2223230237 the review has been approved.

Everything looks good!
