### 2024-08-09 #6857 @kthyng
Congratulations on your new publication @ashjbarnes and @navidcy! Many thanks to @AnjaliSandip and to reviewers @alperaltuntas, @dionhaefner, and @theresa-morrison for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.

@ashjbarnes and @navidcy If you're interested in joining JOSS as a reviewer, please sign up here: https://reviewers.joss.theoj.org/!
