### 2024-08-09 #6857 @navidcy
> @ashjbarnes and @navidcy If you're interested in joining JOSS as a reviewer, please sign up here: https://reviewers.joss.theoj.org/!

Thanks @kthyng!

I think I'm already on the list. But I'm urging the rest of the team to join as well: @aekiss, @dhruvbhagtani, @luweiyang, @ChrisC28, @reillyja
