### 2025-01-23 #6857 @arfon
@ChrisC28 – this should be fixed now. It may take a little while to propagate through the various metadata systems but the correct ORCID has now been assigned/deposited from our side.
